To Version: 0.2.36

+++ Output/0.2.36/Control_Scripts/Initial-Setup.ps1
@@ -79,6 +79,8 @@ if (-not(Test-Path $AnswerFile)) {
 		InstallLicensedApps = $false
 		EnableHyperV        = $false
 		EnableWSL           = $false
+		WSLUser             = 'None'
+		WSLPassword         = 'None'
 	}
 	$output | ConvertTo-Json | Out-File -FilePath $AnswerFile -Force
 	Start-Process -FilePath notepad.exe -ArgumentList $AnswerFile -Wait
@@ -118,7 +120,7 @@ $full = Get-Item "$($PSDownload.FullName)\Start-PSToolkitSystemInitialize.ps1"
 try {
 	Import-Module $full.FullName -Force
 
-	if ($GitHubUserID -like 'None') {Start-PSToolkitSystemInitialize -LabSetup -InstallMyModules}
+	if ($GitHubUserID -like "None") {Start-PSToolkitSystemInitialize -LabSetup -InstallMyModules}
 	else {Start-PSToolkitSystemInitialize -GitHubUserID $GitHubUserID -GitHubToken $GitHubToken -LabSetup -InstallMyModules}
 
 	Remove-Item $full.FullName
@@ -126,9 +128,8 @@ try {
 #endregion
 
 #region baseapps
-if (-not(Test-Path "$($PSDownload.fullname)\BaseApps.tmp") -and ($GitHubUserID -notlike 'None')) {
+if (-not(Test-Path "$($PSDownload.fullname)\BaseApps.tmp") -and ($GitHubUserID -notlike "None")) {
 	try {
-		
 		check-reboot
 		
 		Write-Host "`n`n-----------------------------------" -ForegroundColor DarkCyan; Write-Host '[Installing]: ' -NoNewline -ForegroundColor Yellow; Write-Host 'Base Apps' -ForegroundColor Cyan -NoNewline; Write-Host " (New Window)`n" -ForegroundColor darkYellow
@@ -140,7 +141,7 @@ if (-not(Test-Path "$($PSDownload.fullname)\BaseApps.tmp") -and ($GitHubUserID -
 #endregion
 
 #region all modules
-if ($InstallAllModules -and ($GitHubUserID -notlike 'None')) {
+if ($InstallAllModules -and ($GitHubUserID -notlike "None")) {
 	if (-not(Test-Path "$($PSDownload.fullname)\ExtendedModules.tmp")) {
 		Write-Host "`n`n-----------------------------------" -ForegroundColor DarkCyan; Write-Host '[Installing]: ' -NoNewline -ForegroundColor Yellow; Write-Host 'Extended Modules' -ForegroundColor Cyan -NoNewline; Write-Host " (New Window)`n" -ForegroundColor darkYellow   
 		Start-Process PowerShell -ArgumentList "-NoLogo -NoProfile -WindowStyle Maximized -ExecutionPolicy Bypass -Command (& {Install-PWSHModule -ListName BaseModules, ExtendedModules, MyModules -Scope AllUsers -GitHubUserID $GitHubUserID -GitHubToken $GitHubToken})" -Wait -WorkingDirectory C:\Temp\PSTemp 
@@ -154,9 +155,9 @@ if ($InstallAllApps -and ($GitHubUserID -notlike 'None')) {
 	if (-not(Test-Path "$($PSDownload.fullname)\ExtendedApps.tmp")) {
 		check-reboot
 		Write-Host "`n`n-----------------------------------" -ForegroundColor DarkCyan; Write-Host '[Installing]: ' -NoNewline -ForegroundColor Yellow; Write-Host 'Extended Apps' -ForegroundColor Cyan -NoNewline; Write-Host " (New Window)`n" -ForegroundColor darkYellow   
-		Start-Process PowerShell -ArgumentList "-NoLogo -NoProfile -WindowStyle Maximized -ExecutionPolicy Bypass -Command (& {Install-PSPackageManAppFromList -ListName BaseApps, ExtendedApps -GitHubUserID $GitHubUserID -GitHubToken $GitHubToken})" -Wait -WorkingDirectory C:\Temp\PSTemp 
+		Start-Process PowerShell -ArgumentList "-NoLogo -NoProfile -WindowStyle Maximized -ExecutionPolicy Bypass -Command (& {Install-PSPackageManAppFromList -ListName BaseApps, ExtendedApps -GitHubUserID $GitHubUserID -GitHubToken $GitHubToken})' -Wait -WorkingDirectory C:\Temp\PSTemp 
 		
-		Remove-Item -Path "$([Environment]::GetFolderPath('Desktop'))\*.lnk" -ErrorAction SilentlyContinue
+		Remove-Item -Path '$([Environment]::GetFolderPath('Desktop'))\*.lnk" -ErrorAction SilentlyContinue
 		Remove-Item -Path "$($env:PUBLIC)\Desktop\*.lnk" -ErrorAction SilentlyContinue
 
 		New-Item "$($PSDownload.fullname)\ExtendedApps.tmp" -ItemType file -Force | Out-Null
@@ -181,8 +182,17 @@ if ($EnableHyperV) {
 		Write-Host "`n`n-----------------------------------" -ForegroundColor DarkCyan; Write-Host '[Installing]: ' -NoNewline -ForegroundColor Yellow; Write-Host 'Hyper-V' -ForegroundColor Cyan -NoNewline; Write-Host " (New Window)`n" -ForegroundColor darkYellow   
 		Start-Process PowerShell -ArgumentList '-NoLogo -NoProfile -WindowStyle Maximized -ExecutionPolicy Bypass -Command (& {choco install -y Microsoft-Hyper-V-All --source=windowsFeatures})' -Wait -WorkingDirectory C:\Temp\PSTemp 
 		check-reboot
-		Start-Process PowerShell -ArgumentList "-NoLogo -NoProfile -WindowStyle Maximized -ExecutionPolicy Bypass -Command (& {if (-not(Test-Path C:\Hyper-V)) { New-Item C:\Hyper-V -ItemType Directory -Force | Out-Null };if (-not(Test-Path C:\Hyper-V\VHD)) { New-Item C:\Hyper-V\VHD -ItemType Directory -Force | Out-Null};if (-not(Test-Path C:\Hyper-V\Config)) { New-Item C:\Hyper-V\Config -ItemType Directory -Force | Out-Null};Hyper-V\Set-VMHost -VirtualHardDiskPath 'C:\Hyper-V\VHD' -VirtualMachinePath 'C:\Hyper-V\Config'})" -Wait -WorkingDirectory C:\Temp\PSTemp 
-		Start-Process PowerShell -ArgumentList "-NoLogo -NoProfile -WindowStyle Maximized -ExecutionPolicy Bypass -Command (& {Hyper-V\New-VMSwitch -Name 'External' -NetAdapterName (Get-NetAdapter -Physical | Where-Object {$_.status -like 'up'}).Name -AllowManagementOS $true})" -Wait -WorkingDirectory C:\Temp\PSTemp 
+
+		[scriptblock]$block = {
+			if (-not(Test-Path C:\Hyper-V)) { New-Item C:\Hyper-V -ItemType Directory -Force | Out-Null }
+			if (-not(Test-Path C:\Hyper-V\VHD)) { New-Item C:\Hyper-V\VHD -ItemType Directory -Force | Out-Null}
+			if (-not(Test-Path C:\Hyper-V\Config)) { New-Item C:\Hyper-V\Config -ItemType Directory -Force | Out-Null}
+			Hyper-V\Set-VMHost -VirtualHardDiskPath 'C:\Hyper-V\VHD' -VirtualMachinePath 'C:\Hyper-V\Config'
+		
+			$NetAdap = (Get-NetAdapter -Physical | Where-Object {$_.status -like 'up'})[0]
+			Hyper-V\New-VMSwitch -Name 'External2' -NetAdapterName $NetAdap.Name
+		}
+		Start-Process PowerShell -ArgumentList "-NoLogo -NoProfile -noexit -WindowStyle Maximized -ExecutionPolicy Bypass -Command (& {$($block)} )" -Wait -WorkingDirectory C:\Temp\PSTemp 
 
 		New-Item "$($PSDownload.fullname)\EnableHyperV.tmp" -ItemType file -Force | Out-Null
 	}
@@ -190,29 +200,58 @@ if ($EnableHyperV) {
 #endregion
 
 #region WSL
-if ($EnableWSL) {
+if ($EnableWSL  -and ($WSLUser -notlike "None")) {
 	if (-not(Test-Path "$($PSDownload.fullname)\WSL.tmp")) {
 		check-reboot
 		Write-Host "`n`n-----------------------------------" -ForegroundColor DarkCyan; Write-Host '[Installing]: ' -NoNewline -ForegroundColor Yellow; Write-Host 'WSL' -ForegroundColor Cyan -NoNewline; Write-Host " (New Window)`n" -ForegroundColor darkYellow   
 		Start-Process PowerShell -ArgumentList '-NoLogo -NoProfile -WindowStyle Maximized -ExecutionPolicy Bypass -Command (& {wsl --install})' -Wait -WorkingDirectory C:\Temp\PSTemp 
 		check-reboot
-		
+				
 		[scriptblock]$block = {
-			wsl -d Ubuntu -u root apt update
-			wsl -d Ubuntu -u root apt install make git -y
-			wsl -d Ubuntu -u root git clone "https://$($GitHubToken):x-oauth-basic@github.com/smitpi/ansible-bootstrap" /tmp/ansible/ansible-bootstrap
-			wsl -d Ubuntu -u root cp /tmp/ansible/ansible-bootstrap/inventory-src /tmp/ansible/inventory
-			wsl -d Ubuntu -u root mkdir /tmp/ansible/host_vars
 
-			wsl -d Ubuntu -u root apt install git python3-pip python3-dev -y
-			wsl -d Ubuntu -u root pip3 install ansible
-			wsl -d Ubuntu -u root ansible-galaxy install -r /tmp/ansible/ansible-bootstrap/requirements.yml --force
+			cmd.exe /c "ubuntu run -u root useradd -m -G sudo -s /bin/bash $($WSLUser)"
+			cmd.exe /c "ubuntu run -u root (echo $($WSLPassword); echo $($WSLPassword)) | ubuntu run -u root passwd $($WSLUser)"
+			cmd.exe /c "ubuntu config --default-user $($WSLUser)"
+			cmd.exe /c "ubuntu run -u root echo '$($WSLUser) ALL=(ALL) NOPASSWD:ALL' |  ubuntu run -u root tee /etc/sudoers.d/$($WSLUser)"
 
-			wsl -d Ubuntu -u root ansible-playbook -i /tmp/ansible/inventory /tmp/ansible/ansible-bootstrap/local.yml --limit localhost --tags initial
-			wsl -d Ubuntu -u root ansible-playbook -i /tmp/ansible/inventory /tmp/ansible/ansible-bootstrap/local.yml
-		}		
+			cmd.exe /c 'ubuntu sudo apt update'
+			cmd.exe /c 'ubuntu sudo apt install make git -y'
+			cmd.exe /c "ubuntu git clone https://$($GitHubToken):x-oauth-basic@github.com/smitpi/ansible-bootstrap ~/ansible/ansible-bootstrap"
+			cmd.exe /c 'ubuntu sudo cp ~/ansible/ansible-bootstrap/inventory-src ~/ansible/inventory'
+			cmd.exe /c 'ubuntu sudo mkdir ~/ansible/host_vars'
 
-		Start-Process PowerShell -ArgumentList "-NoLogo -NoProfile -WindowStyle Maximized -ExecutionPolicy Bypass -Command (& {$($block)} )" -Wait -WorkingDirectory C:\Temp\PSTemp 
+			cmd.exe /c 'ubuntu sudo apt install git python3-pip python3-dev -y'
+			cmd.exe /c 'ubuntu sudo pip3 install ansible'
+			cmd.exe /c 'ubuntu ansible-galaxy install -r ~/ansible/ansible-bootstrap/requirements.yml --force'
+
+			cmd.exe /c 'ubuntu ansible-playbook -i ~/ansible/inventory ~/ansible/ansible-bootstrap/local.yml --limit localhost --tags initial'
+			cmd.exe /c 'ubuntu ansible-playbook -i ~/ansible/inventory ~/ansible/ansible-bootstrap/local.yml'
+		}
+
+		[scriptblock]$block1 = {
+			cmd.exe /c 'ubuntu run sudo curl  -o /etc/wsl.conf -L https://raw.githubusercontent.com/smitpi/PSToolKit/master/PSToolKit/Private/Config/wsl.conf'
+			cmd.exe /c 'wsl --terminate Ubuntu'
+		}
+
+		[scriptblock]$block2 = {
+			cmd.exe /c 'ubuntu run sudo apt update'
+
+			cmd.exe /c 'ubuntu run sudo apt install make git -y'
+			cmd.exe /c "ubuntu run git clone https://$($GitHubToken):x-oauth-basic@github.com/smitpi/ansible-bootstrap ~/ansible/ansible-bootstrap"
+			cmd.exe /c 'ubuntu run sudo cp ~/ansible/ansible-bootstrap/inventory-src ~/ansible/inventory'
+			cmd.exe /c 'ubuntu run sudo mkdir ~/ansible/host_vars'
+
+			cmd.exe /c 'ubuntu run sudo apt install git python3-pip python3-dev -y'
+			cmd.exe /c 'ubuntu run sudo pip3 install ansible'
+			cmd.exe /c 'ubuntu run ansible-galaxy install -r ~/ansible/ansible-bootstrap/requirements.yml --force'
+
+			cmd.exe /c 'ubuntu run ansible-playbook -i ~/ansible/inventory ~/ansible/ansible-bootstrap/local.yml --limit localhost --tags initial'
+			cmd.exe /c 'ubuntu run ansible-playbook -i ~/ansible/inventory ~/ansible/ansible-bootstrap/local.yml'
+		}
+
+		Start-Process PowerShell -ArgumentList "-NoLogo -NoProfile -noexit -WindowStyle Maximized -ExecutionPolicy Bypass -Command (& {$($block)} )" -Wait -WorkingDirectory C:\Temp\PSTemp 
+		Start-Process PowerShell -ArgumentList "-NoLogo -NoProfile -noexit -WindowStyle Maximized -ExecutionPolicy Bypass -Command (& {$($block1)} )" -Wait -WorkingDirectory C:\Temp\PSTemp 
+		Start-Process PowerShell -ArgumentList "-NoLogo -NoProfile -noexit -WindowStyle Maximized -ExecutionPolicy Bypass -Command (& {$($block2)} )" -Wait -WorkingDirectory C:\Temp\PSTemp 
 
 		New-Item "$($PSDownload.fullname)\WSL.tmp" -ItemType file -Force | Out-Null
 	}
